samples/ml/ml_jobs/e2e_task_graph/requirements.txt
snowflake-ml-python>=1.9.0
snowflake-snowpark-python
snowflake-core
xgboost==2.1.3